Inital commit of a skeleton project:
This project uses Spring Boot and CXF to host a soap RPC encoded literal API compatable with WITSML.

This skeleton has been confirmed to work with the PDS client for getBaseMsg, getVersion, and getCap. The security has been commented out of the POM until we have a clearer understanding of what we are doing as it relates to security (token/basic).