df-331: Fix search for trajectory stations by md
This fix resolves df #331 Fix search for trajectory stations by md.

With this fix, when a UID is provided by a trajectory request when performing a getObject(), an additional check will be performed. The check will detect if any (or any combination) of mdMn, mdMx, or lastUpdateTimeUtc are present. In this case, the request will be sent via the graphQL path via search (rather than the REST path).