Commit Where Unmarshalling the Response Works
This code check-in is for when the system is providing a response from DoT to the calling client (which is working in WITSML XML). The first step is to correctly unmarshal into the Log POJOs, then perform transformations. One transformation is provided by this check-in. I will complete the transformations (analogous to convertToChannelSet(), only going in the opposite direction).