UpdateInStore:Log channelSet is working COB commit)
Not done!
I still need Junit and to put in Redis.
I am now turning my attention to Redis, as I had a working local Redis example that I am thinking I can glean stuff from prior work I did. In the process of doing this now.
Plus, I verified that the <serviceCompany> changed from Baker Hughes to Halliburton with a GET, but asked Sukhe why I do not see <logData> in the GET response. If I need to verify with a different command, I will do that next. Also the conversion for the response is not in WITSML XML...that may come later down the line, so I asked Sukhe that question also.

I will work on Redis tonight, from home, and hopefully have it in place tomorrow. Once in place, I can do the object comparison & add the code that will not perform the Update unless it is a changed object.