#454 and #460 Fix
This is the fix for classIndex/columnIndex. I am currently blocked in completing verification of the fix due to "getChannels" not correctly returning channels information (currently, I see columnIndex reported in classIndex). I will open a ticket for this as a separate issue & complete testing of this once this is resolved.